requirements.txt
# except native python libraries
PyYAML==6.0
python-box==7.1.1
aioconsole==0.6.2